Describe pair programming in terms to avoid bad practices
It is important that the pair collaborate over the same screen. One
should not be off looking at something else, while the other writes
code. The pair should be seeing the same thing and communicating. When
one member of the pair is looking at something else, the two loose out
on the communication benefits.